requirements.txt
pyjwt~=2.10.1
fastapi~=0.115.12
uvicorn~=0.34.2